Move 404 page component to pages directory
