+++ server/frontend/src/pages/404.tsx
